+++ Assets/Plugins/Source/Editor/Platforms/Android/AndroidBuilder.cs
@@ -344,7 +344,7 @@ namespace PlayEveryWare.EpicOnlineServices.Editor.Build
 
         private static void ConfigureEOSDependentLibrary()
         {
-            string clientIDAsLower = Config.Get<EOSConfig>().clientID.ToLower();
+            string clientIDAsLower = PlatformManager.GetPlatformConfig().clientCredentials.ClientId.ToLower();
 
             var pathToEOSValuesConfig = GetAndroidEOSValuesConfigPath();
             var currentEOSValuesConfigAsXML = new System.Xml.XmlDocument();
