+++ Assets/Plugins/Source/Editor/Platforms/Windows/WindowsBuilder.cs
@@ -91,14 +91,9 @@ namespace PlayEveryWare.EpicOnlineServices.Editor.Build
 #else
             // Determine if 'DisableOverlay' is set in Platform Flags. If it is, then the EOSBootstrapper.exe is not included in the build,
             // because without needing the overlay, the EOSBootstrapper.exe is not useful to users of the plugin
-            if (PlatformManager.TryGetPlatformConfig(out PlatformConfig configuration))
-            {
-                // TODO-URGENT: Better error log
-                Debug.LogError("Could not get platform config.");
-                return;
-            }
+            PlatformConfig platformConfig = PlatformManager.GetPlatformConfig();
 
-            PlatformFlags configuredFlags = configuration.platformOptionsFlags.Unwrap();
+            PlatformFlags configuredFlags = platformConfig.platformOptionsFlags.Unwrap();
             if (configuredFlags.HasFlag(PlatformFlags.DisableOverlay))
             {
                 Debug.Log($"The '{nameof(PlatformFlags.DisableOverlay)}' flag has been configured, EOSBootstrapper.exe will not be included in this build.");
